requirements.txt
django>=5.2.7
pillow
django-crispy-forms
crispy-bootstrap4
scikit-learn
numpy
joblib